Delegate alias generation to macro
To support the multi-arch proposal #166, the extension requires more sophisticated logic for the alias repository, particularly when a single package resolves to multiple RPMs (e.g. for different architectures).

Previously, the repository rule was responsible for the textual construction of alias target in the generated BUILD file. This approach is rigid and scales poorly when complex logic to handle more aliases is required.
By extracting this logic into a dedicated macro, the repository rule is responsible only for providing structured metadata about RPMs. The logic determining how to expose that data (as an alias, a null rule, or perhaps a `select` in the future) is deferred to the macro. This decouples data resolution from target generation and allows utilizing more the expressiveness of Starlark.

This implementation tries to provide the equivalent logic as before.
Minor changes include:
- if an RPM repository was generated for other config, now the alias will be correctly generated (metadata is fetched from previous config via the `registered_rpms` shared dict);
- not creating `null_rpm_repository` as alias generator macro can call `null_rpm_rule` directly.

This structure allows for future customisation, where users could potentially provide their own alias generation strategies if the default one doesn't fit their needs.